## Description
Avoid using jsoniter, move to fastjson

## Motivation and Context
This is to avoid using unsafe.Pointer type
code dependency for MinIO, this causes
crashes on ARM64 platforms

Refer #8005 collection of runtime crashes due
to unsafe.Pointer usage incorrectly. We have
seen issues like this before when using
jsoniter library in the past.

This PR hopes to fix this using fastjson

## How to test this PR?
This is a primary fix for ARM64 crashes but it has added benefits of performance as well 

Running benchmarks here are the differences, `new` is fastjson

```
benchcmp old.txt new.txt 
benchmark                            old ns/op     new ns/op     delta
BenchmarkParseUnmarshal/N10-4        5938          4574          -22.97%
BenchmarkParseUnmarshal/N100-4       41242         37051         -10.16%
BenchmarkParseUnmarshal/N500-4       193704        222566        +14.90%
BenchmarkParseUnmarshal/N1000-4      395749        444372        +12.29%
BenchmarkParseUnmarshal/N5000-4      3128205       2377310       -24.00%
BenchmarkParseUnmarshal/N10000-4     6277548       5672244       -9.64%

benchmark                            old MB/s     new MB/s     speedup
BenchmarkParseUnmarshal/N10-4        503.85       654.05       1.30x
BenchmarkParseUnmarshal/N100-4       657.43       731.79       1.11x
BenchmarkParseUnmarshal/N500-4       699.59       608.87       0.87x
BenchmarkParseUnmarshal/N1000-4      684.82       609.88       0.89x
BenchmarkParseUnmarshal/N5000-4      437.00       575.03       1.32x
BenchmarkParseUnmarshal/N10000-4     436.00       482.53       1.11x

benchmark                            old allocs     new allocs     delta
BenchmarkParseUnmarshal/N10-4        85             64             -24.71%
BenchmarkParseUnmarshal/N100-4       547            514            -6.03%
BenchmarkParseUnmarshal/N500-4       2555           2516           -1.53%
BenchmarkParseUnmarshal/N1000-4      5059           5022           -0.73%
BenchmarkParseUnmarshal/N5000-4      25110          25274          +0.65%
BenchmarkParseUnmarshal/N10000-4     50202          51316          +2.22%

benchmark                            old bytes     new bytes     delta
BenchmarkParseUnmarshal/N10-4        5808          3345          -42.41%
BenchmarkParseUnmarshal/N100-4       44067         28720         -34.83%
BenchmarkParseUnmarshal/N500-4       191586        143420        -25.14%
BenchmarkParseUnmarshal/N1000-4      382152        288887        -24.41%
BenchmarkParseUnmarshal/N5000-4      3651836       1630332       -55.36%
BenchmarkParseUnmarshal/N10000-4     7457806       3807840       -48.94%
```

Run the benchmarks https://gist.github.com/harshavardhana/2ffc990bf3d3ee7ee935f5b146413a72

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
- [ ] Fixes a regression (If yes, please add `commit-id` or `PR #` here)
- [ ] Documentation needed
- [x] Unit tests needed
- [ ] Functional tests needed (If yes, add [mint](https://github.com/minio/mint) PR # here: )