@minio-ops 2019-08-11
## Mint Automation
| Test | Result |
|:--- |:--- |
| mint-compression-fs.sh | :heavy_check_mark: |
| mint-worm.sh | :heavy_check_mark: |
| mint-fs.sh | :heavy_check_mark: |
| mint-gateway-nas.sh | :heavy_check_mark: |
| mint-compression-xl.sh | :x: [more...](#f0794eed-734d-4b3e-bdab-e9c4b279cdd7) |
| mint-xl.sh | :x: [more...](#b357c5f7-6994-4997-a9b0-baeb75e158a0) |
| mint-large-bucket.sh | :x: [more...](#2b8582bd-8791-4094-ac23-72702291c387) |
| mint-compression-dist-xl.sh | :x: [more...](#3f215636-ae4b-477b-a60d-b7d31ed2a884) |
| mint-dist-xl.sh | :x: [more...](#f3615914-7b1e-4ff6-869f-49883ee19be8) |
<a name="f3615914-7b1e-4ff6-869f-49883ee19be8"></a>
### 8063-9d91ad9/mint-dist-xl.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.53:30086
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp d23b9023d56e:/mint/log /tmp/mint-logs'

(1/14) Running aws-sdk-go tests ... done in 11 seconds
(2/14) Running aws-sdk-java tests ... done in 1 seconds
(3/14) Running aws-sdk-php tests ... FAILED in 7 minutes and 28 seconds
{
  "name": "aws-sdk-php",
  "function": "listMultipartUploads ( array $params = [] )",
  "args": {
    "Bucket": "aws-sdk-php-5d50a63ed70a8",
    "Object": "obj1"
  },
  "duration": "444312",
  "status": "FAIL",
  "error": "Error executing \"AbortMultipartUpload\" on \"http://72.28.97.53:30086/aws-sdk-php-5d50a63ed70a8/obj1-copy?uploadId=9cdb98a2-3710-4119-91d5-2f609b22c03d\"; AWS HTTP error: Server error: `DELETE http://72.28.97.53:30086/aws-sdk-php-5d50a63ed70a8/obj1-copy?uploadId=9cdb98a2-3710-4119-91d5-2f609b22c03d` resulted in a `503 Service Unavailable` response:\n<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<Error><Code>XMinioServerNotInitialized</Code><Message>Server not initialized, pl (truncated...)\n XMinioServerNotInitialized (server): Server not initialized, please try again. - <?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<Error><Code>XMinioServerNotInitialized</Code><Message>Server not initialized, please try again.</Message><Key>obj1-copy</Key><BucketName>aws-sdk-php-5d50a63ed70a8</BucketName><Resource>/aws-sdk-php-5d50a63ed70a8/obj1-copy</Resource><RequestId>15BA02DE67D05D1B</RequestId><HostId></HostId></Error>"
}

Executed 2 out of 14 tests successfully.
```
<a name="b357c5f7-6994-4997-a9b0-baeb75e158a0"></a>
### 8063-9d91ad9/mint-xl.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.61:30006
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 808f13d2425a:/mint/log /tmp/mint-logs'

(1/14) Running aws-sdk-go tests ... done in 0 seconds
(2/14) Running aws-sdk-java tests ... done in 2 seconds
(3/14) Running aws-sdk-php tests ... FAILED in 0 seconds
{
  "name": "aws-sdk-php",
  "function": "listMultipartUploads ( array $params = [] )",
  "args": {
    "Bucket": "aws-sdk-php-5d50a63315c1e",
    "Object": "obj1"
  },
  "duration": "630",
  "status": "FAIL",
  "error": "Error executing \"AbortMultipartUpload\" on \"http://72.28.97.61:30006/aws-sdk-php-5d50a63315c1e/obj1-copy?uploadId=79f904a0-aca7-4876-a10a-1865f556b677\"; AWS HTTP error: cURL error 7: Failed to connect to 72.28.97.61 port 30006: Connection refused (see http://curl.haxx.se/libcurl/c/libcurl-errors.html)"
}

Executed 2 out of 14 tests successfully.
```
<a name="f0794eed-734d-4b3e-bdab-e9c4b279cdd7"></a>
### 8063-9d91ad9/mint-compression-xl.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.55:30273
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 0ca77bbd181e:/mint/log /tmp/mint-logs'

(1/14) Running aws-sdk-go tests ... FAILED in 6 minutes and 0 seconds
{
  "alert": "",
  "args": {
    "bucketName": "aws-sdk-go-test-bnw4tc4wg6vul4",
    "expiry": 60000000000,
    "objectName": "presignedTest"
  },
  "duration": 360009,
  "error": "XMinioServerNotInitialized: Server not initialized, please try again.\n\tstatus code: 503, request id: 15BA02C6FFF09E3D, host id: ",
  "function": "PresignedPut",
  "message": "AWS SDK Go CreateBucket Failed",
  "name": "aws-sdk-go",
  "status": "FAIL"
}

Executed 0 out of 14 tests successfully.
```
<a name="3f215636-ae4b-477b-a60d-b7d31ed2a884"></a>
### 8063-9d91ad9/mint-compression-dist-xl.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.60:30200
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 6ed062be10f5:/mint/log /tmp/mint-logs'

(1/14) Running aws-sdk-go tests ... done in 1 seconds
(2/14) Running aws-sdk-java tests ... done in 1 seconds
(3/14) Running aws-sdk-php tests ... FAILED in 7 minutes and 27 seconds
{
  "name": "aws-sdk-php",
  "function": "listMultipartUploads ( array $params = [] )",
  "args": {
    "Bucket": "aws-sdk-php-5d50a6343aace",
    "Object": "obj1"
  },
  "duration": "447025",
  "status": "FAIL",
  "error": "Error executing \"AbortMultipartUpload\" on \"http://72.28.97.60:30200/aws-sdk-php-5d50a6343aace/obj1-copy?uploadId=b1155544-d0f8-4aa5-afd1-0adf5ff5abe4\"; AWS HTTP error: Server error: `DELETE http://72.28.97.60:30200/aws-sdk-php-5d50a6343aace/obj1-copy?uploadId=b1155544-d0f8-4aa5-afd1-0adf5ff5abe4` resulted in a `503 Service Unavailable` response:\n<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<Error><Code>XMinioServerNotInitialized</Code><Message>Server not initialized, pl (truncated...)\n XMinioServerNotInitialized (server): Server not initialized, please try again. - <?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<Error><Code>XMinioServerNotInitialized</Code><Message>Server not initialized, please try again.</Message><Key>obj1-copy</Key><BucketName>aws-sdk-php-5d50a6343aace</BucketName><Resource>/aws-sdk-php-5d50a6343aace/obj1-copy</Resource><RequestId>15BA02DBD55EA750</RequestId><HostId></HostId></Error>"
}

Executed 2 out of 14 tests successfully.
```
<a name="2b8582bd-8791-4094-ac23-72702291c387"></a>
### 8063-9d91ad9/mint-large-bucket.sh.log:
```
Running with
SERVER_ENDPOINT:      72.28.97.61:32319
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 614ed5d18d3e:/mint/log /tmp/mint-logs'

(1/14) Running aws-sdk-go tests ... done in 16 seconds
(2/14) Running aws-sdk-java tests ... done in 2 seconds
(3/14) Running aws-sdk-php tests ... FAILED in 16 minutes and 27 seconds
{
  "name": "aws-sdk-php",
  "function": "listMultipartUploads ( array $params = [] )",
  "args": {
    "Bucket": "aws-sdk-php-5d50a644c0f8d",
    "Object": "obj1"
  },
  "duration": "985740",
  "status": "FAIL",
  "error": "Error executing \"AbortMultipartUpload\" on \"http://72.28.97.61:32319/aws-sdk-php-5d50a644c0f8d/obj1-copy?uploadId=512f75d2-1217-44d5-a0ae-bbe04e44944d\"; AWS HTTP error: cURL error 52: Empty reply from server (see http://curl.haxx.se/libcurl/c/libcurl-errors.html)"
}

Executed 2 out of 14 tests successfully.
```